Refactored code. Added hex class.
