+++ src/client/game_client.py
@@ -1,5 +1,6 @@
 from service import Service
 
+
 class GameClient:
     def __init__(self) -> None:
         pass
