+++ src/client/service.py
@@ -1,3 +1,3 @@
-class Entity:
+class Service:
     def __init__(self) -> None:
         pass
