+++ src/entity/catapult.py
@@ -1,5 +1,6 @@
 from entity import Entity
 
-class Obstacle(Entity):
+
+class Catapult(Entity):
     def __init__(self) -> None:
         super().__init__()
