+++ src/entity/entity.py
@@ -1,0 +1,6 @@
+class Entity:
+    def __init__(self) -> None:
+        pass
+
+    def draw(self) -> None:
+        pass
