+++ src/entity/hard_repair.py
@@ -1,5 +1,6 @@
 from entity import Entity
 
-class Catapult(Entity):
+
+class HardRepair(Entity):
     def __init__(self) -> None:
         super().__init__()
