+++ src/entity/light_repair.py
@@ -1,5 +1,6 @@
 from entity import Entity
 
-class HardRepair(Entity):
+
+class LightRepair(Entity):
     def __init__(self) -> None:
         super().__init__()
