+++ src/entity/obstacle.py
@@ -1,5 +1,6 @@
 from entity import Entity
 
-class LightRepair(Entity):
+
+class Obstacle(Entity):
     def __init__(self) -> None:
         super().__init__()
