+++ src/entity/tank.py
@@ -1,5 +1,6 @@
 from entity import Entity
 
+
 class Tank(Entity):
     def __init__(self) -> None:
         super().__init__()
