+++ src/game.py
@@ -1,8 +1,3 @@
-from player.player import Player
-from game_map import GameMap
-from game_client import GameClient
-from server_enum import *
-
 class Game:
     def __init__(self) -> None:
         pass
