--- src/game/game_map.py
@@ -1,3 +1,0 @@
-class GameMap:
-    def __init__(self):
-        pass
