--- src/game/player/player.py
@@ -1,8 +1,0 @@
-from game import Game
-
-class Player:
-    def __init__(self) -> None:
-        pass
-
-    def playMove(self) -> str:
-        pass
