+++ src/map/game_map.py
@@ -1,0 +1,3 @@
+class GameMap:
+    def __init__(self) -> None:
+        pass
