+++ src/map/hex.py
@@ -1,4 +1,3 @@
-
-class Service:
+class Hex:
     def __init__(self) -> None:
         pass
