+++ src/player/bot_player.py
@@ -1,5 +1,6 @@
 from player import Player
 
+
 class BotPlayer(Player):
     def __init__(self) -> None:
         super().__init__()
