+++ src/player/human_player.py
@@ -1,5 +1,6 @@
 from player import Player
 
+
 class HumanPlayer(Player):
     def __init__(self) -> None:
         super().__init__()
