+++ src/player/player.py
@@ -1,0 +1,9 @@
+from src import game
+
+
+class Player:
+    def __init__(self) -> None:
+        pass
+
+    def play_move(self) -> str:
+        pass
